requirements.txt
numpy>=1.26,<3.0
pandas>=2.2,<3.0
scikit-learn>=1.5,<2.0
matplotlib>=3.9,<4.0
seaborn>=0.13,<1.0
streamlit>=1.43,<2.0
openpyxl>=3.1,<4.0
plotly>=5.24,<7.0
openai>=1.0,<3.0
# ── benchmark pipeline optional dependencies ──────────────────────────────
# ucimlrepo>=0.0.7          # UCI ML Repository loader (pip install ucimlrepo)
# datasets>=2.20,<4.0       # HuggingFace datasets (pip install datasets)
# mlxtend>=0.23,<1.0        # Itemset mining — Apriori/FP-Growth (pip install mlxtend)
# prefixspan>=0.5,<1.0      # Episode mining fallback (pip install prefixspan)
fpdf2>=2.7,<3.0             # PDF report export (pip install fpdf2)
